backend/requirements.txt
requests==2.20.1
python-jose==3.1.0
